# api/openapi/schema.yaml
openapi: 3.0.3
info:
  title: BookHub API
  description: API REST para gerenciamento de empréstimos de livros
  version: 1.0.0
  contact:
    name: BookHub Team
    email: contact@bookhub.com

servers:
  - url: http://localhost:8080/api/v1
    description: Local development server

tags:
  - name: auth
    description: Autenticação
  - name: users
    description: Gerenciamento de usuários
  - name: books
    description: Gerenciamento de livros
  - name: loans
    description: Empréstimos de livros
  - name: nova
    description: Nova categoria de handlers

paths:
  /hello-world:
    get:
      tags:
        - nova
      summary: Exemplo de novo handler
      operationId: my-hello-world
      responses:
        "200":
          description: Retorna mensagem de boas vindas.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HelloWorldResponse"

  /auth/login:
    post:
      tags:
        - auth
      summary: Autenticar usuário
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login bem-sucedido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users:
    get:
      tags:
        - users
      summary: Listar todos os usuários
      operationId: listUsers
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Lista de usuários
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
        "401":
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags:
        - users
      summary: Criar novo usuário
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: Usuário criado com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/{id}:
    get:
      tags:
        - users
      summary: Buscar usuário por ID
      operationId: getUserById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Usuário encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "404":
          description: Usuário não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      tags:
        - users
      summary: Atualizar usuário
      operationId: updateUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Usuário atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Usuário não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/{id}/disable:
    patch:
      tags:
        - users
      summary: Desabilitar usuário
      operationId: disableUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Usuário desabilitado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Usuário não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /books:
    get:
      tags:
        - books
      summary: Listar todos os livros
      operationId: listBooks
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: available
          in: query
          description: Filtrar por disponibilidade
          schema:
            type: boolean
      responses:
        "200":
          description: Lista de livros
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookListResponse"
        "401":
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags:
        - books
      summary: Criar novo livro
      operationId: createBook
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookRequest"
      responses:
        "201":
          description: Livro criado com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookResponse"
        "400":
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /books/{id}:
    get:
      tags:
        - books
      summary: Buscar livro por ID
      operationId: getBookById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Livro encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookResponse"
        "404":
          description: Livro não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /loans:
    get:
      tags:
        - loans
      summary: Listar empréstimos
      operationId: listLoans
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: user_id
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [active, returned]
      responses:
        "200":
          description: Lista de empréstimos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanListResponse"

  /loans/borrow:
    post:
      tags:
        - loans
      summary: Emprestar livro para usuário
      operationId: borrowBook
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BorrowBookRequest"
      responses:
        "201":
          description: Empréstimo realizado com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanResponse"
        "400":
          description: Não é possível realizar empréstimo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Livro ou usuário não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /loans/{id}/return:
    patch:
      tags:
        - loans
      summary: Devolver livro
      operationId: returnBook
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Livro devolvido com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanResponse"
        "400":
          description: Não é possível devolver o livro
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Empréstimo não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HelloWorldResponse:
      type: object
      required:
        - title
      properties:
        title:
          type: string

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: admin@bookhub.com
        password:
          type: string
          format: password
          minLength: 6
          example: "123456"

    LoginResponse:
      type: object
      properties:
        token:
          type: string
        expires_at:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"

    CreateUserRequest:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 100
          example: João Silva
        email:
          type: string
          format: email
          example: joao@email.com
        password:
          type: string
          format: password
          minLength: 6
          example: "senha123"

    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 100
        email:
          type: string
          format: email

    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    UserResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/User"

    UserListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        pagination:
          $ref: "#/components/schemas/Pagination"

    CreateBookRequest:
      type: object
      required:
        - title
        - author
        - isbn
        - total_copies
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: Clean Code
        author:
          type: string
          minLength: 1
          maxLength: 100
          example: Robert C. Martin
        isbn:
          type: string
          pattern: "^[0-9]{10,13}$"
          example: "9780132350884"
        published_year:
          type: integer
          minimum: 1000
          maximum: 2100
          example: 2008
        total_copies:
          type: integer
          minimum: 1
          example: 5

    Book:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        author:
          type: string
        isbn:
          type: string
        published_year:
          type: integer
        total_copies:
          type: integer
        available_copies:
          type: integer
        availability_status:
          type: string
          description: "Disponível ou Indisponível - todas as cópias emprestadas"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    BookResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Book"

    BookListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Book"
        pagination:
          $ref: "#/components/schemas/Pagination"

    BorrowBookRequest:
      type: object
      required:
        - user_id
        - book_id
      properties:
        user_id:
          type: string
          format: uuid
        book_id:
          type: string
          format: uuid
        due_date:
          type: string
          format: date
          description: Data de devolução prevista (padrão 14 dias)

    Loan:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        user_name:
          type: string
        book_id:
          type: string
          format: uuid
        book_title:
          type: string
        borrowed_at:
          type: string
          format: date-time
        due_date:
          type: string
          format: date-time
        returned_at:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [active, returned]

    LoanResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Loan"

    LoanListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Loan"
        pagination:
          $ref: "#/components/schemas/Pagination"

    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer

    MessageResponse:
      type: object
      properties:
        message:
          type: string

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: array
          items:
            type: string