# internal/infrastructure/database/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package sqlc

import (
	"database/sql"
	"time"

	"github.com/google/uuid"
)

type Book struct {
	ID              uuid.UUID     `json:"id"`
	Title           string        `json:"title"`
	Author          string        `json:"author"`
	Isbn            string        `json:"isbn"`
	PublishedYear   sql.NullInt32 `json:"published_year"`
	TotalCopies     int32         `json:"total_copies"`
	AvailableCopies int32         `json:"available_copies"`
	CreatedAt       time.Time     `json:"created_at"`
	UpdatedAt       time.Time     `json:"updated_at"`
}

type Loan struct {
	ID         uuid.UUID    `json:"id"`
	UserID     uuid.UUID    `json:"user_id"`
	BookID     uuid.UUID    `json:"book_id"`
	BorrowedAt time.Time    `json:"borrowed_at"`
	DueDate    time.Time    `json:"due_date"`
	ReturnedAt sql.NullTime `json:"returned_at"`
	Status     string       `json:"status"`
}

type User struct {
	ID           uuid.UUID `json:"id"`
	Name         string    `json:"name"`
	Email        string    `json:"email"`
	PasswordHash string    `json:"password_hash"`
	Active       bool      `json:"active"`
	CreatedAt    time.Time `json:"created_at"`
	UpdatedAt    time.Time `json:"updated_at"`
}